third_party/blink/web_tests/platform/webview/external/wpt/bluetooth/idl/idlharness.tentative.https.window-expected.txt

This is a testharness.js-based test.
Found 160 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Bluetooth interface: existence and properties of interface object
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface object length
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface object name
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: operation getAvailability()
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute onavailabilitychanged
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute referringDevice
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: operation getDevices()
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: operation requestDevice(optional RequestDeviceOptions)
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute onadvertisementreceived
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute ongattserverdisconnected
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute oncharacteristicvaluechanged
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute onserviceadded
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute onservicechanged
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth interface: attribute onserviceremoved
  assert_own_property: self does not have own property "Bluetooth" expected property "Bluetooth" missing
[FAIL] Bluetooth must be primary interface of navigator.bluetooth
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of navigator.bluetooth
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "getAvailability()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onavailabilitychanged" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "referringDevice" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "getDevices()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "requestDevice(optional RequestDeviceOptions)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: calling requestDevice(optional RequestDeviceOptions) on navigator.bluetooth with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onadvertisementreceived" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "ongattserverdisconnected" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "oncharacteristicvaluechanged" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onserviceadded" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onservicechanged" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onserviceremoved" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothPermissionResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface object length
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface object name
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: attribute devices
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] ValueEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface object length
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface object name
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: attribute value
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] BluetoothDevice interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface object length
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface object name
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute id
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute name
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute gatt
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: operation forget()
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: operation watchAdvertisements(optional WatchAdvertisementsOptions)
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute watchingAdvertisements
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute onadvertisementreceived
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute ongattserverdisconnected
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute oncharacteristicvaluechanged
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute onserviceadded
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute onservicechanged
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothDevice interface: attribute onserviceremoved
  assert_own_property: self does not have own property "BluetoothDevice" expected property "BluetoothDevice" missing
[FAIL] BluetoothAdvertisingEvent interface object length
  assert_equals: wrong value for BluetoothAdvertisingEvent.length expected 2 but got 0
[FAIL] BluetoothAdvertisingEvent must be primary interface of event
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of event
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "device" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "uuids" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "name" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "appearance" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "txPower" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "rssi" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "manufacturerData" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "serviceData" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothRemoteGATTServer interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface object length
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface object name
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: attribute device
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: attribute connected
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: operation connect()
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: operation disconnect()
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: operation getPrimaryService(BluetoothServiceUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTServer interface: operation getPrimaryServices(optional BluetoothServiceUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTServer" expected property "BluetoothRemoteGATTServer" missing
[FAIL] BluetoothRemoteGATTService interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface object length
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface object name
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute device
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute uuid
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute isPrimary
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: operation getCharacteristic(BluetoothCharacteristicUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: operation getCharacteristics(optional BluetoothCharacteristicUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: operation getIncludedService(BluetoothServiceUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: operation getIncludedServices(optional BluetoothServiceUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute oncharacteristicvaluechanged
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute onserviceadded
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute onservicechanged
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTService interface: attribute onserviceremoved
  assert_own_property: self does not have own property "BluetoothRemoteGATTService" expected property "BluetoothRemoteGATTService" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface object length
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface object name
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: attribute service
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: attribute uuid
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: attribute properties
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: attribute value
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation getDescriptor(BluetoothDescriptorUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation getDescriptors(optional BluetoothDescriptorUUID)
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation readValue()
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation writeValue(BufferSource)
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation writeValueWithResponse(BufferSource)
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation writeValueWithoutResponse(BufferSource)
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation startNotifications()
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: operation stopNotifications()
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothRemoteGATTCharacteristic interface: attribute oncharacteristicvaluechanged
  assert_own_property: self does not have own property "BluetoothRemoteGATTCharacteristic" expected property "BluetoothRemoteGATTCharacteristic" missing
[FAIL] BluetoothCharacteristicProperties interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface object length
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface object name
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute broadcast
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute read
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute writeWithoutResponse
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute write
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute notify
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute indicate
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute authenticatedSignedWrites
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute reliableWrite
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothCharacteristicProperties interface: attribute writableAuxiliaries
  assert_own_property: self does not have own property "BluetoothCharacteristicProperties" expected property "BluetoothCharacteristicProperties" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface object length
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface object name
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: attribute characteristic
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: attribute uuid
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: attribute value
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: operation readValue()
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothRemoteGATTDescriptor interface: operation writeValue(BufferSource)
  assert_own_property: self does not have own property "BluetoothRemoteGATTDescriptor" expected property "BluetoothRemoteGATTDescriptor" missing
[FAIL] BluetoothUUID interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface object length
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface object name
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: operation getService((DOMString or unsigned long))
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: operation getCharacteristic((DOMString or unsigned long))
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: operation getDescriptor((DOMString or unsigned long))
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] BluetoothUUID interface: operation canonicalUUID(unsigned long)
  assert_own_property: self does not have own property "BluetoothUUID" expected property "BluetoothUUID" missing
[FAIL] Navigator interface: attribute bluetooth
  assert_true: The prototype object must have a property "bluetooth" expected true got false
[FAIL] Navigator interface: navigator must inherit property "bluetooth" with the proper type
  assert_inherits: property "bluetooth" not found in prototype chain
Harness: the test ran to completion.
